tests/submission/W3C/domhtml/HTMLBodyElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-body-background" rel="help">
<link rel="author" title="NIST">
</head>
<body background="/support/images/yellow-rectangle.png"> 
<p>Hello, World</p>


<script>
var domtest = async_test("The background IDL attribute of the body element must reflect the element's background content attribute. (The background content is not defined to contain a URL, despite rules regarding its handling in the rendering section above.)");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vbackground;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("body");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vbackground = testNode.background;

      assert_equals(vbackground,"/support/images/yellow-rectangle.png");
       
    domtest.done();
}));

</script>
<p>The background IDL attribute of the body element must reflect the element's background content attribute. (The background content is not defined to contain a URL, despite rules regarding its handling in the rendering section above.)</p>
<div id="log"></div>
</body></html>
